+++ creature/enemies/skeleton/skeleton.cpp
@@ -1,6 +1,6 @@
 #include "skeleton.h"
 
-Skeleton::Skeleton(CreatureManager& _manager, int _health, const sf::Vector2f& _pos)
+Skeleton::Skeleton(CreatureManager& _manager, float _health, const sf::Vector2f& _pos)
         : Enemy("skeleton", _manager, _health, _pos, {24.f, 44.f}, {14.f, 14.f}, {32.f, 32.f}) {
     creature_type = CreatureType::SKELETON;
     creature_anim = CreatureAnim::HUMANOID;
