+++ creature/enemies/skeleton/skeleton.h
@@ -3,6 +3,6 @@
 
 class Skeleton : public Enemy {
 public:
-    Skeleton(CreatureManager& manager, int health = 100, const sf::Vector2f& pos = {500, 500});
+    Skeleton(CreatureManager& manager, float health = 100.f, const sf::Vector2f& pos = {500, 500});
     virtual Drop drop() const override;
 };
