+++ creature/enemies/spider/spider.cpp
@@ -1,6 +1,6 @@
 #include "spider.h"
 
-Spider::Spider(CreatureManager& _manager, int _health, const sf::Vector2f& _pos)
+Spider::Spider(CreatureManager& _manager, float _health, const sf::Vector2f& _pos)
         : Enemy("spider_low1", _manager, _health, _pos, {32.f, 26.f}, {18.f, 18.f}, {31.f, 16.f}) {
     creature_type = CreatureType::SPIDER;
     creature_anim = CreatureAnim::SPIDER;
