+++ creature/enemies/taur/taur.h
@@ -3,5 +3,5 @@
 
 class Taur : public Enemy {
 public:
-    Taur(CreatureManager& manager, int health = 100, const sf::Vector2f& pos = {128, 128});
+    Taur(CreatureManager& manager, float health = 100.f, const sf::Vector2f& pos = {128, 128});
 };
