+++ creature/enemies/wolf/wolf.h
@@ -3,5 +3,5 @@
 
 class Wolf : public Enemy {
 public:
-    Wolf(CreatureManager& manager, int health = 100, const sf::Vector2f& pos = {128, 128});
+    Wolf(CreatureManager& manager, float health = 100.f, const sf::Vector2f& pos = {128, 128});
 };
